requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pandas==2.0.3
openpyxl==3.1.2
qrcode==7.4.2
Pillow==10.0.0
smtplib-ssl==0.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7